requirements.txt
facebook-sdk==0.4.0
httplib2==0.9
six==1.7.2
twilio==3.6.6
wsgiref==0.1.2
